Use skipBytes(long) in places where integer overflow is possible

+++ components/formats-bsd/src/loci/formats/in/DicomReader.java
@@ -1100,7 +1100,7 @@ public class DicomReader extends FormatReader {
     if (skip) {
       long skipCount = (long) elementLength;
       if (in.getFilePointer() + skipCount <= in.length()) {
-        in.skipBytes((int) skipCount);
+        in.skipBytes(skipCount);
       }
       location += elementLength;
       value = "";
